Updating products

+++ store/products/form.py
@@ -1,9 +1,9 @@
 from flask_wtf.file import FileAllowed, FileField, FileRequired
-from wtforms import Form, IntegerField, StringField, BooleanField, TextAreaField, validators
+from wtforms import Form, IntegerField, StringField, TextAreaField, validators, DecimalField
 
 class Addproducts(Form):
     name = StringField('Nome:', [validators.DataRequired()])
-    price = IntegerField('Preço:', [validators.DataRequired()])
+    price = DecimalField('Preço:', [validators.DataRequired()])
     discount = IntegerField('Desconto:', [validators.DataRequired()])
     stock = IntegerField('Estoque:', [validators.DataRequired()])
     description = TextAreaField('Descrição:', [validators.DataRequired()])
